Refactor record set into a page source provider
