+++ src/main/java/pl/net/was/FakerModule.java
@@ -44,7 +44,7 @@ public class FakerModule
         binder.bind(FakerConnector.class).in(Scopes.SINGLETON);
         binder.bind(FakerMetadata.class).in(Scopes.SINGLETON);
         binder.bind(FakerSplitManager.class).in(Scopes.SINGLETON);
-        binder.bind(FakerRecordSetProvider.class).in(Scopes.SINGLETON);
+        binder.bind(FakerPageSourceProvider.class).in(Scopes.SINGLETON);
         binder.bind(FakerPageSinkProvider.class).in(Scopes.SINGLETON);
         configBinder(binder).bindConfig(FakerConfig.class);
     }
